requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.3.1
alt-profanity-check==1.1.3
async-timeout==4.0.2
attrs==21.2.0
black==22.12.0
Brotli==1.0.9
cchardet==2.1.7
certifi==2022.12.7
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
et-xmlfile==1.1.0
frozenlist==1.3.3
function-cooldowns==1.6.0
idna==3.3
joblib==1.2.0
motor==2.5.1
multidict==5.2.0
mypy-extensions==0.4.3
nextcord==2.3.2
nextcord-ext-menus==1.3.2
numpy==1.24.0
openai==0.25.0
openpyxl==3.0.10
orjson==3.6.4
pandas==1.5.2
pandas-stubs==1.5.2.221213
pathspec==0.10.3
platformdirs==2.6.0
pycares==4.1.2
pycparser==2.20
pymongo==3.12.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
tabulate==0.9.0
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
urllib3==1.26.13
validators==0.20.0
yarl==1.7.0
